Move MakeGlobalSettingsWithFile to a separate file from TestingEnvironment.php.
